LayoutTests/platform/mac-wk1/accessibility/custom-elements/current-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL accessibilityController.accessibleElementById("custom-location").currentStateValue should be location (of type string). Was undefined (of type undefined).
FAIL accessibilityController.accessibleElementById("custom-page").currentStateValue should be page (of type string). Was undefined (of type undefined).
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
